requirements.txt
numpy 
pillow 
tensorflow 
scikit-learn 
xgboost 
matplotlib 
seaborn 
scikit-plot
jupyterlab